+++ Includes/Brotli.h
@@ -1,6 +1,6 @@
 #ifndef AEM_INCLUDES_BROTLI_H
 #define AEM_INCLUDES_BROTLI_H
 
-int brotliCompress(char** holder, int* lenData);
+int brotliCompress(char **holder, size_t *lenData);
 
 #endif
